Add script to download weather data from OpenGribs API
This commit adds a Python script to download weather data from the OpenGribs API. The script sends a request with a sample URL to the API and parses the response to get the URL to download. It then downloads the data in GRIB2 format and saves it in a file named "test.grb2". The commit also includes a .gitignore file to ignore GRIB2 files in the repository.